+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * feat(scm-repo-link): Add `SCMRepoLinkList` method [#241](https://github.com/Scalingo/go-scalingo/pull/241) and [#245](https://github.com/Scalingo/go-scalingo/pull/245)
 * build(deps): bump github.com/golang-jwt/jwt/v4 from 4.1.0 to 4.4.1
 * build(deps): bump github.com/stretchr/testify from 1.7.0 to 1.7.1
+* feat(log-drains): The `LogDrainsAddonList` now returns the list of log drains [#246](https://github.com/Scalingo/go-scalingo/pull/246)
+* feat(log-drains): Cleanup the LogDrain struct [#246](https://github.com/Scalingo/go-scalingo/pull/246)
 
 ## 4.15.1